+++ mzident/src/formats/mztab.rs
@@ -1045,7 +1045,7 @@ fn parse_modification<'a>(
             let modification = if tag.eq_ignore_ascii_case("unimod") {
                 ontologies
                     .unimod()
-                    .get_by_index(&value.parse::<usize>().map_err(|err| {
+                    .get_by_index(&value.parse::<u32>().map_err(|err| {
                         BoxedError::new(
                             BasicKind::Error,
                             "Invalid unimod code",
@@ -1064,7 +1064,7 @@ fn parse_modification<'a>(
             } else if tag.eq_ignore_ascii_case("mod") {
                 ontologies
                     .psimod()
-                    .get_by_index(&value.parse::<usize>().map_err(|err| {
+                    .get_by_index(&value.parse::<u32>().map_err(|err| {
                         BoxedError::new(
                             BasicKind::Error,
                             "Invalid PSI-MOD code",
@@ -1083,7 +1083,7 @@ fn parse_modification<'a>(
             } else if tag.eq_ignore_ascii_case("custom") {
                 ontologies
                     .custom()
-                    .get_by_index(&value.parse::<usize>().map_err(|err| {
+                    .get_by_index(&value.parse::<u32>().map_err(|err| {
                         BoxedError::new(
                             BasicKind::Error,
                             "Invalid custom code",
